refactor: structure all cli commands, options
improves approachability of the code as a whole. no significant functional changes, just reorganisation
