+++ src/about/cli.ts
@@ -7,28 +7,26 @@ import { Command } from 'commander'
 /**
  * Display community/further reading links in CLI.
  */
-const communityAction = () => async () => {
-  console.log([
-    'Browse public Edge code on GitHub:\n',
-    '  https://github.com/edge\n\n',
-    'Chat with the community on Discord, Telegram, or Reddit:\n',
-    '  https://discord.gg/edge-network\n',
-    '  https://t.me/edgenetwork\n',
-    '  https://reddit.com/r/edgenetwork\n\n',
-    'Follow Edge on Twitter or Facebook:\n',
-    '  https://twitter.com/edgenetwork\n',
-    '  https://www.facebook.com/edgenetworktech\n\n',
-    'Keep up with Edge on Medium or sign up to the mailing list:\n',
-    '  https://edge.medium.com\n',
-    '  https://edge.press'
-  ].join(''))
+const community = async () => {
+  console.log(`
+Browse public Edge code on GitHub:
+  https://github.com/edge
+
+Chat with the community on Discord, Telegram, or Reddit:
+  https://discord.gg/edge-network
+  https://t.me/edgenetwork
+  https://reddit.com/r/edgenetwork
+
+Follow Edge on Twitter or Facebook:
+  https://twitter.com/edgenetwork
+  https://www.facebook.com/edgenetworktech
+
+Keep up with Edge on Medium or sign up to the mailing list:
+  https://edge.medium.com
+  https://edge.press
+`)
 }
 
-/** Configure about commands. */
-export const commands = (): Command[] => {
-  const community = new Command('community')
-    .description('show community links')
-    .action(communityAction())
-
-  return [community]
-}
+export const commands = (): Command[] => [
+  new Command('community').description('show community links').action(community)
+]
