--- src/edge/color.ts
@@ -1,18 +1,0 @@
-// Copyright (C) 2021 Edge Network Technologies Limited
-// Use of this source code is governed by a GNU GPL-style license
-// that can be found in the LICENSE.md file. All rights reserved.
-
-import chalk from 'chalk'
-
-/**
- * Standard colours for log output.
- */
-export default {
-  context: chalk.gray,
-  debug: chalk.blue,
-  error: chalk.red,
-  info: chalk.white,
-  notice: chalk.green,
-  success: chalk.green,
-  warn: chalk.yellow
-}
